+++ src/Actions/Statistics/TakeQueueSnapshot.php
@@ -4,7 +4,6 @@ namespace VincentBean\HorizonDashboard\Actions\Statistics;
 
 use Illuminate\Database\Eloquent\Builder;
 use Laravel\Horizon\Contracts\MetricsRepository;
-use VincentBean\HorizonDashboard\Http\Controllers\GetCpuMemoryUsage;
 use VincentBean\HorizonDashboard\Models\JobStatistic;
 use VincentBean\HorizonDashboard\Models\QueueStatistic;
